[infra-q8-nhwc] Fix packaging, clang-format, and CodeQL CI failures
Install conv_indirect.h so the Android NDK build and Tizen/Debian
packaging find the header that compute_ops.h includes; it was listed
in the spec and debian .install manifests but missing from the meson
cpu_backend_headers install set, so every packaging job exited with
a missing-file error and the Android build failed with
'conv_indirect.h' file not found.

Also fix a clang-format indentation drift in conv_indirect.h that the
cpp-linter flagged, and widen two unsigned-int multiplication chains
in the NHWC slice/upsample fast-paths so they evaluate in size_t
before use as an index or pointer offset, clearing the CodeQL
integer-overflow alert. The casts are value-preserving; no behavior
changes.

+++ nntrainer/layers/slice_layer.cpp
@@ -129,7 +129,7 @@ static void sliceForwardT(const Tensor &input, Tensor &output,
     } else if (axis == 1) {
       auto &tm = ThreadManager::Global();
       for (unsigned int b = 0; b < B; ++b) {
-        tm.parallel_for(0, Ho * Wo, [&](size_t hw) {
+        tm.parallel_for(0, static_cast<size_t>(Ho) * Wo, [&](size_t hw) {
           std::memcpy(out + ((size_t)b * Ho * Wo + hw) * Co,
                       in + ((size_t)b * Hi * Wi + hw) * Ci + start, Co * elt);
         });
